fix error when connecting to network with no name defined
