+++ packages/safe-apps-developer-ui/src/api/provider.ts
@@ -19,20 +19,6 @@ const WALLET_PROVIDER = {
   UNKNOWN: 'UNKNOWN',
 };
 
-enum ETHEREUM_NETWORK_TO_ID {
-  MAINNET = 1,
-  MORDEN = 2,
-  ROPSTEN = 3,
-  RINKEBY = 4,
-  GOERLI = 5,
-  KOVAN = 42,
-  XDAI = 100,
-  ENERGY_WEB_CHAIN = 246,
-  VOLTA = 73799,
-  UNKNOWN = 0,
-  LOCAL = 4447,
-}
-
 const providerOptions: IProviderOptions = {
   walletconnect: {
     package: WalletConnectProvider,
@@ -50,4 +36,4 @@ const connectToProvider = async (): Promise<any> => {
   return provider;
 };
 
-export { WALLET_PROVIDER, ETHEREUM_NETWORK_TO_ID, connectToProvider };
+export { WALLET_PROVIDER, connectToProvider };
