+++ packages/safe-apps-developer-ui/src/components/Header/components/NetworkLabel.tsx
@@ -3,8 +3,8 @@ import { ReactElement } from 'react';
 import Typography from '@material-ui/core/Typography';
 
 import { border, md, screenSm, sm, xs, fontColor } from 'src/styles/variables';
-import { ETHEREUM_NETWORK_TO_ID } from 'src/api/provider';
 import { upperFirst } from 'src/utils/strings';
+import { getNetworkNameById } from 'src/api/eth';
 
 const useStyles = makeStyles({
   container: {
@@ -39,7 +39,7 @@ const NetworkLabel = ({ networkId }: Props): ReactElement => {
   return (
     <div className={classes.container}>
       <Typography className={classes.text} variant="body2">
-        {upperFirst(ETHEREUM_NETWORK_TO_ID[networkId])}
+        {upperFirst(getNetworkNameById(networkId))}
       </Typography>
     </div>
   );
