+++ packages/safe-apps-developer-ui/src/stores/contracts.ts
@@ -1,5 +1,4 @@
 import create from 'zustand';
-import { ETHEREUM_NETWORK_TO_ID } from 'src/api/provider';
 import { getFromLocalStorage, saveToStorage } from 'src/utils/localStorage';
 
 type ContractNames = 'fallbackHandler' | 'proxyFactory' | 'masterCopy' | 'multiSend';
@@ -9,14 +8,14 @@ type ContractsMap = Record<string, DeployedContracts>;
 
 type ContractsState = {
   contracts: ContractsMap;
-  saveContracts: (networkId: ETHEREUM_NETWORK_TO_ID, deployedContracts: DeployedContracts) => void;
+  saveContracts: (networkId: number, deployedContracts: DeployedContracts) => void;
 };
 
 const CONTRACTS_STORE_KEY = 'contracts_state_v1';
 
 const useContractsStore = create<ContractsState>((set, get) => ({
   contracts: getFromLocalStorage(CONTRACTS_STORE_KEY) || {},
-  saveContracts(networkId: ETHEREUM_NETWORK_TO_ID, deployedContracts: DeployedContracts) {
+  saveContracts(networkId: number, deployedContracts: DeployedContracts) {
     const { contracts } = get();
 
     const newContracts = { ...contracts, [networkId]: deployedContracts };
